backend/requirements.txt
foxglove-sdk>=0.16.0